Added roleable module for User model
